# documentation/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Snippy API",
    "version": "0.9.2",
    "description": "API documentation for Snippy"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token for authorization"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "SnippetList": {
        "type": "object",
        "properties": {
          "snippetId": {
            "type": "string"
          },
          "shortId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "userName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "commentCount": {
            "type": "integer"
          },
          "favoriteCount": {
            "type": "integer"
          },
          "viewCount": {
            "type": "integer"
          },
          "embedCount": {
            "type": "integer"
          },
          "isOwner": {
            "type": "boolean"
          },
          "isFavorited": {
            "type": "boolean"
          },
          "isFollowing": {
            "type": "boolean"
          },
          "parentShortId": {
            "type": "string",
            "nullable": true
          },
          "parentName": {
            "type": "string",
            "nullable": true
          },
          "parentUserName": {
            "type": "string",
            "nullable": true
          },
          "snapshotUrl": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "SnippetListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "snippets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnippetList"
            }
          },
          "totalCount": {
            "type": "integer"
          }
        }
      },
...
      "FavoriteResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "isFavorited": {
            "type": "boolean"
          },
          "favoriteCount": {
            "type": "integer"
          }
        }
      },
      "FavoriteStatusResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "isFavorited": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/favorites/{snippetId}": {
      "post": {
        "operationId": "toggleFavorite",
        "tags": [
          "Favorites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "snippetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Toggled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getFavoriteStatus",
        "tags": [
          "Favorites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "snippetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/favorites": {
      "get": {
        "operationId": "getFavorites",
        "tags": [
          "Favorites"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnippetListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}